Final product, minor features
Added the ability to see and change the map size from NewGame using the arrow keys. Added the ability to see the full map by pressing M. Fixed a bug which caused the minimap to stretch on small maps. Added a debug output which writes every 1000 tries of generating a map.